requirements.txt
requests
python-dotenv
openai
nomeroff-net
fastapi==0.119.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg[binary]==3.2.3
pydantic==2.12.2
pydantic-settings==2.5.2
email-validator==2.2.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
opencv-python==4.12.0.88
numpy==2.1.3
Pillow==11.0.0
ultralytics==8.3.214
torch==2.5.1
torchvision==0.20.1
alembic==1.13.3
